Initial commit of working version
This is the initial commit of a bitmask.h file containing the macros and types required, and main.cpp with statically checked tests and sample usage. There will be some refactoring and additional operations need to be added, as well as some type-safe way of allowing access to the raw underlying type (for things like serialization).